Move tzdb over to timezone_provider::tzif, remove old neri_schneider
